Version 0.0.15
* Added startMySessionV2
* Added more tests
* Removed invalid fields from `UserAuthResponse`

+++ src/operations/myUser/deleteMyUser.ts
@@ -17,6 +17,10 @@ const deleteMyUser = async (
   const clientInfo = clientInfoStore.get();
   const signedOutUserId = clientInfo.myUserId || clientInfo.signedOutUserId;
 
+  if (!clientInfo.myUserId) {
+    throw new Error('not-authorized');
+  }
+
   try {
     await fsdata.myUser.deleteMyUser(cause, description, deletePhysically);
